THERE IS NO GOD
But seriously, this commit brings all sorts of things. The document now
requires titlepic, multirow, and hyperref for the title picture, shiny
tables, and proper hyperlinks/bookmarking respectively. I have taken the
liberty of fixing the numerous syntax and logic errors that were causing
compiling issues. Additionally I added the Domain Model section and
spell checked the rest of the document. Tabs or spaces; please pick one
and be consistent...